moved all tests into a test directory
