+++ tests/auth-routerPG.test.js
@@ -2,8 +2,8 @@ const express = require("express");
 const server = express();
 const request = require("supertest");
 const bcrypt = require("bcryptjs");
-const Users = require("./auth-modal");
-const authRouter = require("./auth-routerPG");
+const Users = require("../auth/auth-modal");
+const authRouter = require("../auth/auth-routerPG");
 // jest.mock("./auth-modal");
 server.use(express.json());
 server.use("/", authRouter);
