+++ tests/server.test.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const testServer = express();
 const request = require("supertest");
-const server = require("./server.js");
+const server = require("../api/server.js");
 testServer.use(express.json());
 testServer.use("/", server);
